+++ src/UI/words.ts
@@ -59,6 +59,24 @@ const dict = {
   "error.server.intro": (errorMessage: string) =>
     `The following error occured while communicating with the server: ${errorMessage}`,
   "error.authorizationFailed": "Authorization failed, please log in",
+  true: "True",
+  false: "False",
+  submit: "Submit",
+  confirm: "Confirm",
+  "inventory.addInstance.title": (serviceName: string) =>
+    `Create a new instance of ${serviceName} with the following parameters`,
+  "inventory.addInstance.button": "Add instance",
+  "inventory.editInstance.button": "Edit",
+  "inventory.editInstance.title": "Edit instance",
+  "inventory.editInstance.header": (instanceId: string) =>
+    `Change attributes of instance ${instanceId}`,
+  "inventory.deleteInstance.button": "Delete",
+  "inventory.deleteInstance.title": "Delete instance",
+  "inventory.deleteInstance.header": (
+    instanceId: string,
+    serviceName: string
+  ) =>
+    `Are you sure you want to delete instance ${instanceId} of service entity ${serviceName}?`,
   "error.environment.missing": "Environment is missing",
   error: "Something went wrong",
   loading: "Loading",