+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Show resource deployment progress in the inventory view (#154)
 - Add auto-updating of data to ResourcesTab (#225)
 - Add auto-updating of service instances to ServiceInventory (#225)
+- Add separate create instance page (#181)
 
 # Release 1.4.0 (2020-12-17)